feat: create bottom panel component
