+++ src/renderer/components/_molecules/variablePanel/index.tsx
